// types/FaqOptionsType.tsx
export type FaqOptionsType = {
  showFaqCategoriesAsGrid?: boolean;
  showContactUsOnFaqScreens?: boolean;
  showContactUsOnAppBar?: boolean;
  showContactUsOnFaqNotHelpful?: boolean;
  tags?: string[] | null;
  contactusFilterTags?: string[] | null;
  contactusFilterTitle?: string | null;
  filteredViewTitle?: string | null;
  filterType?: string | null;
};